+++ install_zsh.sh
@@ -2,15 +2,25 @@
 
 echo_blue "installing zsh"
 
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+#sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 
+if [ -f ~/.zsh_aliases ]; then
+    rm ~/.zsh_aliases
+fi
 ln -s ~/scripts/dotfiles/zsh_aliases ~/.zsh_aliases
-rm ~/.zshrc
+
+if [ -f ~/.zshrc ]; then
+    rm ~/.zshrc
+fi
 ln -s ~/scripts/dotfiles/zshrc ~/.zshrc
+
+if [ -f ~/.oh-my-zsh/themes/mat.zsh-theme ]; then
+    rm ~/.oh-my-zsh/themes/mat.zsh-theme
+fi
 ln -s ~/scripts/dotfiles/mat.zsh-theme ~/.oh-my-zsh/themes/mat.zsh-theme
 source ~/.zshrc
 
 echo_green "zsh installed"
 
-chsh -s /bin/zsh
-zsh
+#chsh -s /bin/zsh
+#zsh
